+++ bot/bot.go
@@ -81,13 +81,13 @@ type FileMetadata struct {
 	LenBytes int64  `json:"length_bytes"`
 }
 
-func (ub *UndercastBot) handleError(ctx context.Context, chatID int, err error) {
+func (ub *UndercastBot) handleError(ctx context.Context, chatID int64, err error) {
 	id := uuid.New().String()
 	ub.logger.Error("error", zap.String("id", id), zap.Error(err))
 	ub.sendTextMessage(ctx, chatID, "An error occurred while processing your request. Please try again later. \nError ID: %s", id)
 }
 
-func (ub *UndercastBot) sendTextMessage(ctx context.Context, chatID int, message string, args ...interface{}) {
+func (ub *UndercastBot) sendTextMessage(ctx context.Context, chatID int64, message string, args ...interface{}) {
 	if _, err := ub.bot.SendMessage(ctx, &bot.SendMessageParams{
 		ChatID: chatID,
 		Text:   fmt.Sprintf(message, args...),
