+++ bot/unpublish_handler.go
@@ -29,7 +29,7 @@ or for range of episodes
 
 func (ub *UndercastBot) unpublishEpisodesHandler(ctx context.Context, b *bot.Bot, update *models.Update) {
 	zapFields := []zap.Field{
-		zap.Int("chatID", update.Message.Chat.ID),
+		zap.Int64("chatID", update.Message.Chat.ID),
 		zap.String("messageText", update.Message.Text),
 	}
